refactor: change api directory structure
