+++ frontend/components/applicant/Board.component.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Board from "@/components/common/board/Board.component";
-import { getApplicantByPage } from "@/src/apis/applicant/applicant";
+import { getApplicantByPage } from "@/src/apis/applicant";
 import ApplicantDetailRight from "./DetailRight.component";
 import ApplicantDetailLeft from "./DetailLeft.component";
 import { useState } from "react";
