+++ frontend/components/applicant/applicantNode/ShortSplit.component.tsx
@@ -1,5 +1,5 @@
 import Txt from "@/components/common/Txt.component";
-import { ApplicantReq } from "@/src/apis/applicant/applicant";
+import { ApplicantReq } from "@/src/apis/applicant";
 import { applicantDataFinder } from "@/src/functions/finder";
 
 interface ApplicantShortSplitProps {
