+++ frontend/src/apis/interview/index.ts
@@ -18,12 +18,14 @@ export interface interviewReqBody {
   record: string;
 }
 
+/** THIS API MAY UNUSED. PLEASE REMOVE THIS COMMENT IF YOU WANT USE */
 export const postInterviewRecord = async (recode: interviewReqBody) => {
   const { data } = await https.post<interviewReqBody>(`/records`, recode);
 
   return data;
 };
 
+/** THIS API MAY UNUSED. PLEASE REMOVE THIS COMMENT IF YOU WANT USE */
 export const getInterviewRecordAll = async () => {
   const { data } = await https.get<InterviewRes[]>(`/records/all`);
 
@@ -65,6 +67,35 @@ export const putInterviewer = async ({ id, role }: putInterviewerReq) => {
   const { data } = await https.put<string>(`/interviewers/${id}/roles`, null, {
     params: { role },
   });
+
+  return data;
+};
+
+export interface putInterviewReq {
+  applicantId: string;
+  record?: string;
+  url?: string;
+}
+
+export const putInterviewRecord = async ({
+  applicantId,
+  record,
+}: putInterviewReq) => {
+  const { data } = await https.put<string>(
+    `/applicants/${applicantId}/records`,
+    { record }
+  );
+  return data;
+};
+
+export const putInterviewUrl = async ({
+  applicantId,
+  url,
+}: putInterviewReq) => {
+  const { data } = await https.put<string>(
+    `/applicants/${applicantId}/records`,
+    { url }
+  );
 
   return data;
 };
